requirements.txt
numpy==1.26.4
wandb==0.19.10

#Huggingface
python-dotenv==1.1.0
transformers==4.50.3
pillow==11.1.0
#flash-attn==2.7.4.post1

#Langchain
langchain==0.3.22
langchain-experimental==0.3.4
unstructured[md]
sentence-transformers==4.0.1
chromadb==0.6.3
langchain-openai==0.3.11

#Unsloth
unsloth==2025.3.19
xformers==0.0.29.post2
trl==0.15.2
peft==0.15.1
triton==3.2.0
accelerate==1.5.2
bitsandbytes==0.45.4
vllm==0.8.2

#ROS stuff
roslibpy==1.8.1

#misc
matplotlib==3.10.1
